requirements.txt
requests
ruamel.yaml
pyyaml